program: Remove build.rs and use frozen-abi feature
#### Problem

As described in https://github.com/anza-xyz/agave/pull/2911, the
frozen-abi bits are feature gated, so we can simply enable the
`min_specialization` feature based on the frozen-abi feature, which is
only ever activated during nightly builds.

#### Summary of changes

Remove build.rs, and use the frozen-abi feature. Note that once we
update to use the 2.1 crates, we'll need to enable the `frozen-abi`
feature on the frozen-abi crates, as done in the previously mentioned
PR.