deps: Upgrade to Solana v2.3.4 and component crates (#63)
* deps: Upgrade to Solana v2.3.4 and component crates

#### Problem

The component crates of the sdk and v2.3.4 of Agave are out, but the
address-lookup-table program is still on old versions for all of these.

#### Summary of changes

Update the rust-toolchain, Agave version, Agave crates, sdk crates, and
component crates.

Also update codama to a newer version to avoid clippy errors, and add
the `fetch` feature.

Update bench numbers accordingly.

* Update rust client versions

* Update lockfile

* Update compute unit numbers

* Wait a bit longer for program to deploy before testing